Make match tests more robust
